fix: productModelCustomizer for localized model (#109)
While testing #108, I found out that the productModelCustomizer isn't working properly for v2 jobs.
Indeed, the `createAlgoliaCategoryObject` function is expecting to have a non-localized model
(i.e. have an object for the `name` property, containing all locales). For v2 jobs, as the `name` property
is a string and not an object, it generates the following output:

```
{
    level_0: { '0': 'N', '1': 'e', '2': 'w', ...
    level_1: { '0': 'N > E', '1': 'e > l', '2': 'w > e', ...
}
```

### Changes

- Add a `createAlgoliaLocalizedCategoryObject` function that directly reads the `name` properties
- Add tests