fix: system object definitions (#118)
System Object types definitions have been broken in #110
Trying to import them result in the following error:

> ERROR [144|33] cvc-complex-type.2.4.d: Invalid content was found
starting with element 'mandatory-flag'. No child element is expected at
this point.

It turns out that the `default-value` is apparently expected at the end.

Also quick question: why has it been marked as mandatory in the
definition? It's not marked as mandatory in the UI, and since it's a
number, it defaults to 0.0 in any case I think?