refactor: compute base-product specific fields in the product model (#145)
In preparation of model extensibility, we need `algoliaLocalizedProduct`
to be the only place that compute the fields, so it will be easier to extend.

### Changes

- Move `variants`, `defaultVariantID` and `colorVariations` computation
in the `algoliaLocalizedProduct` model.
- Removed the `generateMasterRecords()` helper
  - The jobs now generate one `AlgoliaLocalizedProduct` for each locales
instead of calling the `generateMasterRecords()` helper
  - Note: there was some small optimization to pre-fetch a partial variant
model with non-localized fields (such as price) and use it for all
locales. This can't be moved inside the `algoliaLocalizedProduct` model
as it works with a single locale. But it doesn't degrade significantly
the performances, and it's needed if we want to make the model extensible.
- `AlgoliaLocalizedProduct`: removed the requirement to register
attributes in `algoliaProductConfig.attributeConfig_v2`: if the
attribute is present in `aggregatedValueHandlers`, it is computed.
- Removed the `in_stock` parameter at the master-level, this will be reworked later

---
SFCC-267