feat: multiple pricebooks (#147)
## Multiple pricebooks
### Indexing

Add a new `pricebooks` field declaration. When used, it will add all
prices from available pricebooks to the records, for each enabled
currency:

```json
{
  "pricebooks": {
    "USD": {
      "price": 10,
      "pricebookID": "sale-prices",
      "onlineFrom": 0,
      "onlineTo": 3155760000000,
    }
  }
}
```

### Frontend
When present in the records, the default UI will use it to display the strikeout price:
Our default UI display a strikeout price when we have a `price` and a
`promotionalPrice`. If we have a price higher than the actual `price` in the pricebooks

**note**: This only works for Variation Level record model. We indeed
don't have a way to configure which fields should go in the `variants`
objects for the "Base Level" record model currently.

---
SFCC-265