fix(categories): split batches by locale (#171)
The categories job was sending one Algolia batch for each top-level
category, with all locales:
We were iterating only once on each top level category, and fetching all
sub-categories of all locales.

Top-level categories can contain a big amount of sub-categories.
Combined with a big amount of locales, this could lead to a batch too
big to fit in SFCC limitations (20000 objects in an array).

This PR inverts the loops. We will now:
- Iterate on each locale
- For each locale, iterate on each top-level category and build a
complete batch for this locale
- Send the batch and continue with the next locale

This permits the category job to support up to 20000 categories.

### How to test

Run the categories job in a new index.
Index content should be identical to an index generated with the previous version.