chore: use action-semantic-pull-request to validate PR title (#205)
We use the Conventional Commits syntax in most other places as it forces
us to use a meaningful title and make it easier to draft the release
notes. Let's enforce it on our repo.

This PR adds a job to our CI workflow, which uses
[action-semantic-pull-request](https://github.com/amannn/action-semantic-pull-request)
to check the PR title.