feat: allow partial update for product level (#234)
The Base product record model stores all variation attributes in a
`variants` array.

Algolia doesn't permits to update a specific properties in the objects
of an array, so the `variants` array need to be recomputed entirely.
To prevent mistakes, we had initially chosen to block the
`partialRecordUpdate` for the base product model.

But there are cases where users need to only update a few top level
attributes of product records, for example if the records are updated
from multiple indexing pipelines.

This PR removes this safeguard.

It also stop sending the default attributes when `attributeListOverride` is set.