refactor: main jobs logic (#235)
This PR refactors the main jobs `process` function to make it easier to maintain.
**It only reorganizes the code, no logic is changed except a couple of ifs merged.

Some of the logic to generate **variant** records is currently mixed up
in the `if (paramRecordModel === MASTER_LEVEL` block.
This was because when variants have some `attributesComputedFromBaseProduct`,
we generate them all at once from the **base** product, so it sounded initially a good idea
to put the code in the "MASTER_LEVEL" block.

But it actually make it more difficult to read, and makes it harder to add new record models.

The PR refactors the logic to have:
```js
if (paramRecordModel === MASTER_LEVEL) {
  // generate base-product records
} else if (attributesComputedFromBaseProduct.length > 0) {
    // generate variant-level records computed from the base product
}
```

---
SFCC-452