feat: variation group level indexing (#236)
This PR implements Variation Group level indexing for the
`algoliaProductIndex` job.

In this first version, the grouping is done on the `color` attribute
because it's the most common one. We can introduce an option to
configure it if there is a need in the future.
It doesn't rely on SFCC's Variation Groups because they require to
be explicitly defined on products.
The implemented solution uses the variation attributes directly.

### Changes
#### Generation of the records

The main logic lives in the `generateVariationGroupRecords` function:
- For a given `variationAttributeId`, we loop on all the available values
- For each value, we `setSelectedAttributeValue` on the variation model
- We build an Algolia record (`AlgoliaLocalizedProduct`) using this
  variation model. The record is similar to a Base-Product record, with
  the following exceptions:
    - the `variants` array contains only the variant of the selected color
    - the `color` attribute is set at the root level
    - the `objectID` is `<baseProduct.ID>-<colorAttributeValue.ID>`

The `algoliaProductIndex` job calls this function.
If a product doesn't have a `color` variation attribute, then we generate
variant records, using the existing `generateVariantRecords` helper.

---
SFCC-452
SFCC-453
SFCC-454