feat: variation group level inventory hook (#238)
Update our `inventoryUpdate` hook to take into account the
`variation-group-level` record model introduced in #236

The main change is in the `createProductConfig` logic where we set the
current variation attribute value to the base product variation model,
to be able to rebuild the whole `variants` list.

The rest is similar to how Base products are handled.

---
SFCC-456