feat: [SFCC-466] Falling back to master record shape for masters that don't have the specified variation attribute (#240)
The original intention was to fall back to one variation per record for
masters that don't have the specified variation attribute (currently
"color" is hardcoded, will be changed later).
This creates inconsistencies -- falling back to a record of the same
shape as the variation group-level records is a better fit, this PR aims
to do that.

Features:
- masters whose variants that can't be grouped by the specified
variation attribute are now indexed as masters, with the `variants`
array containing all its variants
- updated `algoliaProductIndex.js`
- updated  `algoliaProductDeltaIndex.js`
- records have the same shape no matter which job creates them (tested),
variant deletion and record deletion also work with the delta job if any
variant or all variants go out of stock, respectively (also tested),
going back into stock when there's stock also works (tested).
- the variation group-level records originally had a `color` attribute
hoisted up to root level. If we make the variation attribute
configurable in a subsequent PR, this is going to cause inconsistencies
and confusion, moved it back to the variants in the `variants` array.
- added a lot of explanatory comments to the job flows so that they're
easier to follow and understand