feat: [SFCC-466] Fallback for simple products for the VARIATION_GROUP_LEVEL record model (#242)
In order to stay consistent with masters and records grouped by a
variation attribute, simple products now have a `variants` array when
using the `VARIATION_GROUP_LEVEL` record model, just like masters do.
The `variants` array contains the same attributes that a master's
variant would have.
This change was required in order for the general shape of the records
to be the same and to make faceting work with both masters and simple
products alike.

* updated `algoliaProductIndex.js` job script to handle simple products
* updated `algoliaProductDeltaIndex.js` job script to handle simple
products
* updated `algoliaLocalizedProduct.js` with the new model
* created new helper method `isSimpleProduct` in `jobHelper.js`
* updated the frontend Algolia component to handle the changes
* created new unit tests and updated existing ones
* updated JSDoc for some methods
* added more explanatory comments to existing logic